step-4-google-docs-skill/requirements.txt
google-api-python-client==2.186.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.3
